[MASTERDIET-32] Add achievement verification system.

+++ src/main/java/pl/agh/edu/master_diet/core/model/rest/AchievementsResponse.java
@@ -27,8 +27,8 @@ public class AchievementsResponse {
     private Integer points;
 
     @EqualsAndHashCode.Include()
-    private Integer completeCondition;
+    private Float completeCondition;
     private String imageUrl;
-    private Integer progress;
+    private Float progress;
     private LocalDateTime completedDate;
 }
